compile sous linux

+++ includes/sh_hci.h
@@ -23,7 +23,7 @@ char	*sh_hist_able(char *esc, t_line *line, int *hist_search);
 int		sh_tab(t_line *line, t_coord **coord, t_tc tc);
 int		sh_tab_ins(t_line *line, t_coord **coord, t_tc tc, char *buf);
 char	**sh_tab_find(char *array[], char *dir, int cat);
-int		sh_tab_init(char *str, size_t cur, char **lexeme, int *status);
+int		sh_tab_init(char *str, int cur, char **lexeme, int *status);
 int		sh_tab_multi(t_line *line, t_coord **coord, t_tc tc, char *array[]);
 
 /*
